+++ NEWS.md
@@ -1,6 +1,11 @@
 ## Next 0.6.0
 
 * Added method Position2d#set_pose
+* Added alternate method for subscribing:
+
+  ```ruby
+    robot.ranger(1) # eql robot.subscribe(:ranger, index: 1)
+  ```
 
 ## 2012-04-13 0.5.1